requirements.txt
httpx~=0.27.0
playwright~=1.41.2
pytz~=2024.1
SQLAlchemy~=2.0.28
lxml~=5.2.2
feedparser~=6.0.11
python-dotenv~=1.0.1
redis~=5.0.3
